server/llm-inference-service/requirements.txt
fastapi
uvicorn[standard]
weaviate-client>=4.0.0
langchain
langchain-openai
langchain-community
pandas
numpy
openai
python-dotenv
pydantic
psycopg2-binary
sqlalchemy